backend/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
djangorestframework-simplejwt==4.7.2
Pillow==8.4.0